[mono] Allow overriding GetCustomAttributesData routines
* Merge common bits of Mono and CoreCLR CustomAttributeData implementation

* Have CustomAttributeData.GetCustomAttributes call target.GetCustomAttributesData
  to allow derived classes to override the latter (like in CoreCLR)

* Use Constructor, ConstructorArguments, and NamedArguments in internal
  routines to allow them being overridden (like with CoreCLR).
