+++ src/mono/System.Private.CoreLib/src/System/Reflection/RuntimeAssembly.cs
@@ -275,7 +275,7 @@ namespace System.Reflection
 
         public override IList<CustomAttributeData> GetCustomAttributesData()
         {
-            return CustomAttributeData.GetCustomAttributes(this);
+            return CustomAttributeData.GetCustomAttributesInternal(this);
         }
 
         public override object[] GetCustomAttributes(bool inherit)
