+++ src/mono/System.Private.CoreLib/src/System/Reflection/RuntimeEventInfo.cs
@@ -201,7 +201,7 @@ namespace System.Reflection
 
         public override IList<CustomAttributeData> GetCustomAttributesData()
         {
-            return CustomAttributeData.GetCustomAttributes(this);
+            return CustomAttributeData.GetCustomAttributesInternal(this);
         }
 
         public override int MetadataToken
