+++ src/mono/System.Private.CoreLib/src/System/Reflection/RuntimeFieldInfo.cs
@@ -269,7 +269,7 @@ namespace System.Reflection
 
         public override IList<CustomAttributeData> GetCustomAttributesData()
         {
-            return CustomAttributeData.GetCustomAttributes(this);
+            return CustomAttributeData.GetCustomAttributesInternal(this);
         }
 
         private void CheckGeneric()
