+++ src/mono/System.Private.CoreLib/src/System/Reflection/RuntimeModule.cs
@@ -341,7 +341,7 @@ namespace System.Reflection
 
         public override IList<CustomAttributeData> GetCustomAttributesData()
         {
-            return CustomAttributeData.GetCustomAttributes(this);
+            return CustomAttributeData.GetCustomAttributesInternal(this);
         }
 
         internal RuntimeAssembly GetRuntimeAssembly()
