+++ src/mono/System.Private.CoreLib/src/System/Reflection/RuntimeParameterInfo.cs
@@ -206,7 +206,7 @@ namespace System.Reflection
 
         public override IList<CustomAttributeData> GetCustomAttributesData()
         {
-            return CustomAttributeData.GetCustomAttributes(this);
+            return CustomAttributeData.GetCustomAttributesInternal(this);
         }
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
